+++ core/src/com/trongthien/zbattle/controller/combat/hitbox/HitBox.java
@@ -6,8 +6,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class HitBox {
-    private float x;
-    private float y;
+    private int x;
+    private int y;
     private int width;
     private int height;
